No register keyword in Windows builds.
The documentation states it does nothing (besides introducing warnings in C++).
If it did affect the ABI, my sense is that that would be bad.
Perhaps there should be a separate define such as CMSREGISTER_INTERNAL for use on compilers where the register keyword does make a difference so that shared libraries can disable it in their ABI, but it can be used internally.  Or perhaps using CMS_DLL and CMS_DLL_BUILD should force using register off, rather than on.